Use CreateSlice to combine reducer and actions - makes slice.js for CreateSlice - initailState and reducer in redcuer move to slice.js - reducer's payload is fixed - adds CreateSlice prefix name : application
